Moved everything to src folder to clearn it up.
