+++ Server/gulpfile.js
@@ -11,23 +11,14 @@ gulp.task('watch', ['build'], function(){
   gulp.watch(files, ['build', 'reload'] )
 })
 
-var files = [
-  './api/**/*.js',
-  './config/*.js',
-  './methods/*.js',
-  './Models/*.js',
-  './Schemas/*.js',
-  './Utils/*.js',
-  './tests/**/*.js',
-  './Matching/**/*.js',
-  './app.js'
-]
+var files = './src/**/*.js'
+
 gulp.task('build', function(cb){
   while(children.length != 0){
     children.pop().kill()
   }
 
-  gulp.src(files, { base: './' })
+  gulp.src(files, { base: './src' })
     .pipe(babel())
     .pipe(gulp.dest('dist'))
     .on('end', cb)
