+++ Server/src/Models/GhostTeam.js
