+++ Server/tests/tests.js
@@ -1,4 +1,4 @@
-global.server = require('../app.js');
+global.server = require('../dist/app.js');
 global.chai = require('chai');
 global.expect = global.chai.expect;
